# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Woowup-assessment API",
        "version": "0.1"
    },
    "paths": {
        "/send-mail-failover": {
            "post": {
                "tags": [
                    "Mail"
                ],
                "operationId": "sendMailFailOver",
                "requestBody": {
                    "description": "Request body for sending mail",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "text",
                                    "emailAddresses",
                                    "g-recaptcha-response"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the email",
                                        "type": "string"
                                    },
                                    "text": {
                                        "description": "Text content of the email",
                                        "type": "string"
                                    },
                                    "emailAddresses": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "g-recaptcha-response": {
                                        "description": "Google reCAPTCHA validation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully sent mail"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    }
}